ci: pin go version and use latest toolchain
Starting in Go 1.21, the Go distribution consists of a go command and a
bundled Go toolchain, which is the standard library as well as the
compiler, assembler, and other tools. The go command can use its bundled
Go toolchain as well as other versions that it finds in the local PATH
or downloads as needed.

Signed-off-by: Michael Adler <michael.adler@siemens.com>